+++ packages/design-system/src/components/OcBreadcrumb/types.ts
@@ -1,0 +1,8 @@
+import { RouteLocationRaw } from 'vue-router'
+
+export interface BreadcrumbItem {
+  text: string
+  to?: RouteLocationRaw
+  allowContextActions?: boolean
+  onClick?: () => void
+}
